Re-refactored page names
Still wasn't working after creating new branch, completely renamed files to avoid lowercase overlap
